made movementgroup its own file
